+++ kotlin-compiler-extensions/main/kotlin/com/javiersc/kotlin/compiler/extensions/ir/context/receivers/IrGeneration.kt
@@ -1,4 +1,4 @@
-package com.javiersc.kotlin.compiler.extensions.ir
+package com.javiersc.kotlin.compiler.extensions.ir.context.receivers
 
 import org.jetbrains.kotlin.backend.common.extensions.IrGenerationExtension
 import org.jetbrains.kotlin.backend.common.extensions.IrPluginContext
@@ -9,7 +9,7 @@ public fun interface IrGeneration {
     context(IrModuleFragment, IrPluginContext)
     public fun generate()
 
-    public val extension: IrGenerationExtension get() = object : IrGenerationExtension {
+    public fun getExtension(): IrGenerationExtension = object : IrGenerationExtension {
         override fun generate(moduleFragment: IrModuleFragment, pluginContext: IrPluginContext) {
             with(moduleFragment) {
                 with(pluginContext) {