+++ kotlin-compiler-extensions/main/kotlin/com/javiersc/kotlin/compiler/extensions/ir/irTypeArgument.kt
@@ -1,0 +1,7 @@
+package com.javiersc.kotlin.compiler.extensions.ir
+
+import org.jetbrains.kotlin.ir.types.IrTypeArgument
+import org.jetbrains.kotlin.ir.util.dumpKotlinLike
+
+public fun Iterable<IrTypeArgument?>.dumpKotlinLike(): String =
+    joinToString("\n") { "${it?.dumpKotlinLike()}" }