+++ kotlin-compiler-extensions/main/kotlin/com/javiersc/kotlin/compiler/extensions/shared/as.kt
@@ -1,5 +1,0 @@
-package com.javiersc.kotlin.compiler.extensions.shared
-
-import org.jetbrains.kotlin.ir.IrElement
-
-public inline fun <reified I : IrElement> IrElement?.asIr(): I? = this@asIr as? I?