+++ kotlin-compiler-test-extensions/main/kotlin/com/javiersc/kotlin/compiler/test/runners/DiagnosticTest.kt
@@ -2,8 +2,11 @@ package com.javiersc.kotlin.compiler.test.runners
 
 // import org.jetbrains.kotlin.test.FirParser
 // import org.jetbrains.kotlin.test.directives.configureFirParser
+
+import com.javiersc.kotlin.compiler.test.services.MetaRuntimeClasspathProvider
 import org.jetbrains.kotlin.compiler.plugin.CompilerPluginRegistrar.ExtensionStorage
 import org.jetbrains.kotlin.config.CompilerConfiguration
+import org.jetbrains.kotlin.test.Constructor
 import org.jetbrains.kotlin.test.builders.TestConfigurationBuilder
 import org.jetbrains.kotlin.test.model.TestModule
 import org.jetbrains.kotlin.test.services.EnvironmentBasedStandardLibrariesPathProvider
@@ -11,7 +14,7 @@ import org.jetbrains.kotlin.test.services.KotlinStandardLibrariesPathProvider
 
 public abstract class DiagnosticTest : BaseTestRunner() {
 
-    public val shouldUseAnnotations: Boolean = false
+    public open val runtimeClasspathProvider: Constructor<MetaRuntimeClasspathProvider>? = null
 
     public abstract fun ExtensionStorage.registerExtensions(
         module: TestModule,
@@ -19,7 +22,7 @@ public abstract class DiagnosticTest : BaseTestRunner() {
     )
 
     override fun TestConfigurationBuilder.configuration() {
-        commonFirWithPluginFrontendConfiguration(shouldUseAnnotations) { module, configuration ->
+        commonPluginConfiguration(runtimeClasspathProvider) { module, configuration ->
             registerExtensions(module, configuration)
         }
         //        configureFirParser(FirParser.Psi)