+++ CHANGELOG.md
@@ -4,7 +4,30 @@
 
 ### Added
 
+- ir `annotations` extensions
+- ir `callableId` extensions
+- ir `fqName` extensions
+- ir `irCall` extensions
+- ir `irClass` extensions
+- ir `irClassId` extensions
+- ir `irClassId` extensions
+- ir `irConst` extensions
+- ir `irElement` extensions
+- ir `irFunction` extensions
+- ir `irGeneration` extensions
+- ir `irGetEnumValue` extensions
+- ir `irReturn` extensions
+- ir `irReturn` extensions
+- ir `irTreeNode` extensions
+- ir `irType` extensions
+- ir `irTypeArgument` extensions
+- ir `irValueParameter` extensions
+- ir `name` extensions
+- ir `aIr` extensions
+
 ### Changed
+
+- `kotlin-compiler-test` to `kotlin-compiler-extensions-test`
 
 ### Deprecated